Move a bunch of stuff around to match existing patterns
